# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ManyChat Clone — Instagram DM Automation
  version: 0.1.0
  description: >
    Personal Instagram DM automation. The webhook endpoints are called by
    Meta; everything under /api/admin is the local cockpit and assumes the
    app is bound to localhost (no auth layer — do not expose it publicly).
servers:
  - url: http://localhost:3000
    description: Local Docker
...
paths:
  /api/webhook/instagram:
    get:
      tags: [webhook]
      summary: Subscription handshake
      description: >
        Meta calls this once when you save the callback URL. Echoes
        hub.challenge when hub.verify_token matches IG_VERIFY_TOKEN.
      parameters:
        - { in: query, name: hub.mode, schema: { type: string }, example: subscribe }
        - { in: query, name: hub.verify_token, schema: { type: string } }
        - { in: query, name: hub.challenge, schema: { type: string } }
      responses:
        "200":
          description: Challenge echoed
          content:
            text/plain:
              schema: { type: string }
        "403": { description: Token mismatch }
    post:
      tags: [webhook]
      summary: Receive messages and comments
      description: >
        Verified with HMAC-SHA256 over the raw body. Returns 200 immediately
        and processes asynchronously — Meta retries anything slower than
        ~20s. Deliveries are deduped by message id / comment id.
      parameters:
        - in: header
          name: x-hub-signature-256
          required: true
          schema: { type: string }
          example: sha256=8f2a...
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MetaWebhookPayload" }
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties: { received: { type: boolean } }
        "401": { description: Invalid or missing signature }
...
components:
  schemas:
...
    MetaWebhookPayload:
      type: object
      properties:
        object: { type: string, example: instagram }
        entry:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              messaging:
                type: array
                items:
                  type: object
                  properties:
                    sender:
                      type: object
                      properties: { id: { type: string, description: IGSID } }
                    message:
                      type: object
                      properties:
                        mid: { type: string }
                        text: { type: string }
                        is_echo: { type: boolean }
              changes:
                type: array
                items:
                  type: object
                  properties:
                    field: { type: string, example: comments }
                    value:
                      type: object
                      properties:
                        id: { type: string }
                        text: { type: string }
                        media:
                          type: object
                          properties: { id: { type: string } }
                        from:
                          type: object
                          properties:
                            id: { type: string }
                            username: { type: string }